refactor(client): merge App and main
