+++ packages/client/src/main.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 
-import { App } from './App';
+import { App } from './main';
 
-describe('<App/>', () => {
+describe('<App />', () => {
   it('should render Home page at launch', async () => {
     // Given
     render(<App />);
